% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_common_all.R
\name{build_numeric_selectImput}
\alias{build_numeric_selectImput}
\title{build_numeric_selectImput}
\usage{
build_numeric_selectImput(
  df,
  input_id,
  label_id,
  selected_item,
  add_items = c("none")
)
}
\arguments{
\item{df}{A dataframe}

\item{input_id}{A string}

\item{label_id}{A string}

\item{selected_item}{A string}

\item{add_items}{A vector}
}
\value{
A combobox filled with the names of the numeric columns of the dataframe
}
\description{
build_numeric_selectImput
}
