% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_common_all.R
\name{fill_palette_selector}
\alias{fill_palette_selector}
\title{fill_palette_selector}
\usage{
fill_palette_selector(input_id = "cbPaletteSelector")
}
\arguments{
\item{input_id}{}
}
\value{

}
\description{
fill_palette_selector
}
