% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_common_all.R
\name{fill_plant_selection}
\alias{fill_plant_selection}
\title{fill_plant_selection}
\usage{
fill_plant_selection(
  df,
  input_id = "cbPlantSelection",
  label_id = "Select plants to be displayed",
  selected_text_format = "count > 3",
  preseselct_all = T
)
}
\arguments{
\item{df}{A dataframe}

\item{input_id}{A widget's Id}

\item{label_id}{The widget's label}

\item{selected_text_format}{Format to use when selecting test}

\item{preseselct_all}{Initialize with all options selected?}
}
\value{

}
\description{
fill_plant_selection
}
