% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_common_all.R
\name{fill_time_selection}
\alias{fill_time_selection}
\title{fill_time_selection}
\usage{
fill_time_selection(
  df,
  input_id = "cbDateTimeSelector",
  label_id = "Select days to be displayed"
)
}
\arguments{
\item{df}{A dataframe}

\item{input_id}{A widget's Id}

\item{label_id}{The widget's label}
}
\value{

}
\description{
fill_time_selection
}
