% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_common_all.R
\name{fill_treatment_selection}
\alias{fill_treatment_selection}
\title{fill_treatment_selection}
\usage{
fill_treatment_selection(
  df,
  input_id = "cbTreatmentSelection",
  label_id = "Select treatments to be displayed",
  selected_text_format = "count > 3"
)
}
\arguments{
\item{df}{A dataframe}

\item{input_id}{A widget's Id}

\item{label_id}{The widget's label}

\item{selected_text_format}{Format to use when selecting test}
}
\value{

}
\description{
fill_treatment_selection
}
